+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.0.16 - 2022-11-13
+## v0.0.16 - 2022-11-14
 
 ### Fixed
 
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 
 - Show until when a quiz is silenced in the progress table.
+- Add listening-only quizzes. Closes [#43](https://github.com/fniessink/toisto/issues/43).
 
 ## v0.0.15 - 2022-11-08